# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/fizzbuzz/top": {
            "get": {
                "description": "Returns usage statistics of the /api/v1/fizzbuzz endpoint.\nIt allows the users to know what the number of hits of that endpoint.\nAnd returns the parameters corresponding to it.",
                "produces": [
                    "application/json"
                ],
                "summary": "Most used /api/v1/fizzbuzz request usage statistics.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FizzBuzzTopResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.FizzBuzzParams": {
            "type": "object",
            "properties": {
                "int1": {
                    "description": "all multiples of [Int1] are replaced by [Str1] (classic fizzbuzz: 3)",
                    "type": "integer"
                },
                "int2": {
                    "description": "all multiples of [Int2] are replaced by [Str2] (classic fizzbuzz: 5)",
                    "type": "integer"
                },
                "limit": {
                    "description": "limit of the loop",
                    "type": "integer"
                },
                "str1": {
                    "description": "replacement string for multiples of Int1 (classic fizzbuzz: \"fizz\")",
                    "type": "string"
                },
                "str2": {
                    "description": "replacement string for multiples of [Int2] (classic fizzbuzz: \"buzz\")",
                    "type": "string"
                }
            }
        },
        "service.ErrorMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "service.FizzBuzzTopResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/service.FizzBuzzTopResponseData"
                }
            }
        },
        "service.FizzBuzzTopResponseData": {
            "type": "object",
            "properties": {
                "count_request": {
                    "type": "integer"
                },
                "params": {
                    "$ref": "#/definitions/core.FizzBuzzParams"
                }
            }
        },
...
    }
}